charger/twc3: split fleet code into twc3_fleet.go and extract newTwc3Fleet
Move the Tesla Fleet parts (twc3Fleet, command payloads, schedule/get_config
calls, region/DIN/energy-site discovery, jq queries) into a new twc3_fleet.go,
leaving twc3.go as the core vitals-based charger. The inline Fleet setup in
NewTwc3FromConfig becomes newTwc3Fleet, so the constructor reads as "core
charger + optional fleet block". Pure refactor, no behavior change.
